% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_pseudo_alignment.R
\name{binning}
\alias{binning}
\title{binning}
\usage{
binning(x = x, type = "ACGT")
}
\arguments{
\item{x}{Positive numerical vector.}

\item{type}{If \code{ACGT}, zeros are assigned to T and the rest of the positive values are uniformly binned and assigned accordingly.}
}
\value{
Character vector with entries T, G, C, A.
}
\description{
Maps a positive, continuous vector to discrete states T, G, C, A.
}
