% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_sc_network.R
\name{cluster_heatmap}
\alias{cluster_heatmap}
\title{cluster_heatmap}
\usage{
cluster_heatmap(
  obj,
  distmat,
  group_by = "cluster_names",
  col_by = NULL,
  col_pal = NULL,
  sum_stats = c("min", "mean")[1],
  hierarchical_clustering = FALSE,
  fixed_order = NULL,
  cl_method = c("average", "ward.D", "ward.D2", "single", "complete", "mcquitty",
    "median", "centroid")[1],
  show_plots = T,
  save_plots = F,
  return_matrix = F,
  pre = "",
  dir_plot = getwd(),
  save_height = 4,
  save_width = 4.6,
  ...
)
}
\arguments{
\item{obj}{Seurat object.}

\item{distmat}{Distance matrix.}

\item{group_by}{Name of metadata column to group cells by.}

\item{col_by}{Name of metadata column to color cells by.}

\item{col_pal}{Color palette whose names match the unique values of \code{col_by}.}

\item{sum_stats}{Summary statistics for heatmap entries if \code{group_by} groups multiple cells into one category.}

\item{hierarchical_clustering}{If \code{TRUE} columns and rows of the heatmap are hierarchically clustered.}

\item{fixed_order}{Vector of characters indicating a fixed order of heatmap columns and rows. If \code{NULL} the heatmap columns and rows are not ordered.}

\item{cl_method}{Clustering method for the hierarchical clustering of heatmap columns and rows (see \code{pheatmap::pheatmap()} for details).}

\item{show_plots}{boolean}

\item{save_plots}{boolean}

\item{pre}{Prefix used for name of created PNG.}

\item{dir_plot}{Directory in which PNG file is saved.}

\item{...}{Additional parameters passed to \code{pheatmap::pheatmap()}.}

\item{im_width}{Plot width.}

\item{im_height}{Plot height.}
}
\description{
Wrapper of \code{pheatmap::pheatmap()} visualizing \code{distmat} values as a heatmap and saving it as PNG.
}
