% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_sc_network.R
\name{edge_weight_length_plot}
\alias{edge_weight_length_plot}
\title{edge_weight_length_plot}
\usage{
edge_weight_length_plot(
  obj,
  network,
  group_by = "cluster_names",
  col_by = "cluster_names",
  col_pal = NULL,
  x_attr_name = "min_length",
  y_attr_name = "weight",
  size_attr_name = "min_branch_count",
  pt_maxsize = 2,
  pt_stroke = 0,
  pt_alpha = 0.3,
  show_plots = T,
  save_plots = F,
  pre = "",
  dir_plot = getwd(),
  format = c("png", "pdf")[1],
  im_width = 3,
  im_height = 2.5
)
}
\arguments{
\item{obj}{Seurat object.}

\item{network}{\code{igraph} object.}

\item{group_by}{Name of metadata column to group cells by.}

\item{col_by}{Name of metadata column to color cells by.}

\item{col_pal}{Color palette whose names match the unique values of \code{col_by}.}

\item{x_attr_name}{Edge attribute of \code{network} plotted on x-axis.}

\item{y_attr_name}{Edge attribute of \code{network} plotted on y-axis.}

\item{size_attr_name}{Edge attribute of \code{network} determining size of points.}

\item{pt_maxsize}{Maximal point size.}

\item{pt_stroke}{Passed to \code{ggplot2::geom_point()}.}

\item{pt_alpha}{Passed to \code{ggplot2::geom_point()}.}

\item{show_plots}{boolean}

\item{save_plots}{boolean}

\item{pre}{Prefix used for name of created plot file.}

\item{dir_plot}{Directory in which plot file is saved.}

\item{format}{File format.}

\item{im_width}{Plot width.}

\item{im_height}{Plot height.}

\item{...}{Additional parameters passed to \code{ggplot2::geom_point()}.}
}
\description{
Plots edge lengths vs edge weights (or other attribute combinations) of a network.
}
