% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_sc_network.R
\name{embed_network}
\alias{embed_network}
\title{embed_network}
\usage{
embed_network(
  obj,
  network,
  umap_key = "network_umap",
  edge_length_attr = c("min_length", "min_branch_count")[1],
  min_dist_umap = 0.3,
  n_neighbors_umap = 30,
  spread_umap = 1,
  seed_umap = 42
)
}
\arguments{
\item{obj}{Seurat object.}

\item{network}{\code{igraph} object.}

\item{umap_key}{Name of dimensionality reduction to use in the Seurat object.}

\item{edge_length_attr}{Edge attribute of \code{network} to use for computation of the shortest path distance between cells. If \code{constant} constant edge lengths are used.}

\item{min_dist_umap}{Passed to \code{uwot::umap()}.}

\item{n_neighbors_umap}{Passed to \code{uwot::umap()}.}

\item{spread_umap}{Passed to \code{uwot::umap()}.}

\item{seed_umap}{Passed to \code{uwot::umap()}.}
}
\value{
Seurat object in which the resulting UMAP is stored in \code{obj@reductions[[umap_key]]} and the
the shortest path distance matrix is stored in \code{obj@misc$cellREST_distmat}.
}
\description{
Embeds the nodes of a network based on edge lengths using UMAP.
}
