% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_plotting.R
\name{guide_small_legend_cont}
\alias{guide_small_legend_cont}
\title{guide_small_legend_cont}
\usage{
guide_small_legend_cont()
}
\description{
Uses \code{ggplot2::guides()}, i.e. size and position, for plotting continuous \code{fill} and \code{color} values.
}
