% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_plotting.R
\name{guide_small_legend_disc}
\alias{guide_small_legend_disc}
\title{guide_small_legend_disc}
\usage{
guide_small_legend_disc(...)
}
\description{
Adapts \code{ggplot2::guides()}, i.e. size and position, for plotting discrete \code{fill} and \code{color} values.
}
