% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_import_from_IQ_Tree.R
\name{import_cpu_time}
\alias{import_cpu_time}
\title{import_cpu_time}
\usage{
import_cpu_time(dir)
}
\arguments{
\item{dir}{Directory, in which of \code{iqtree_parallel.log} file is stored.}
}
\value{
Numeric vector with CPU times for tree reconstructions.
}
\description{
Reads combined CPU time from \code{iqtree_parallel.log} file which recorded multiple tree reconstructions.
}
