% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_import_from_IQ_Tree.R
\name{import_trees_lh}
\alias{import_trees_lh}
\title{import_trees_lh}
\usage{
import_trees_lh(dir)
}
\arguments{
\item{dir}{Directory in which IQ-TREE output files \code{.runtrees} or \code{.treefile} and \code{.log} are stored.}
}
\value{
List containing trees as \code{multiPhylo} object and a numeric vector of likelihoods.
}
\description{
Imports trees and likelihoods available in \code{.runtrees} files or multiple \code{.treefile} and \code{.log} files.
}
