% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_sc_network.R
\name{knn_tree}
\alias{knn_tree}
\title{knn_tree}
\usage{
knn_tree(
  obj,
  tree_index = 1,
  tree_list = tree_list,
  pre_comp_path_dists_mat = NULL,
  tree = NULL,
  k = NULL,
  k_dist = 1,
  branch_counts = T,
  return_graph = F
)
}
\arguments{
\item{obj}{Seurat object.}

\item{tree_index}{Index of tree in list of trees \code{tree_list}.}

\item{tree_list}{List of trees or \code{multiPhylo} object.}

\item{pre_comp_path_dists_mat}{Pre-computed distances output by \code{plot_tree_UMAP()} (optional).}

\item{tree}{Phylogenetic tree of class \code{phylo}. If \code{NULL}, tree in \code{tree_list[[tree_index]]} is used.}

\item{k}{Number of neighbors for the nearest neighbor graph construction. If \code{NULL} the maximal neighbor distance in \code{k_dist} is considered.}

\item{k_dist}{Maximal distance for the nearest neighbor graph construction, if \code{k = NULL}.}

\item{branch_counts}{If \code{TRUE} branch counts are used for the evaluation of the maximal distance \code{k_dist}, if \code{FALSE}, path lengths are used.}

\item{return_graph}{If \code{TRUE} the \code{igraph} object is returned, if \code{FALSE} a list of edges is returned.}
}
\description{
Creates a nn-tree-graph from a phylogenetic tree using either \code{k} nearest neighbors or all neighbors closer than \code{k_dist}. The phylogenetic tree can be either input directly via \code{tree}, or a via specifiying a list of trees \code{tree_list} and
\code{tree_index}. Optionally, pre-computed patristic distances \code{pre_comp_path_dists_mat} output by \code{plot_tree_UMAP()} can be input to save computational time.
}
