% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_import_from_IQ_Tree.R
\name{likelihoods_from_logs}
\alias{likelihoods_from_logs}
\title{likelihoods_from_logs}
\usage{
likelihoods_from_logs(dir)
}
\arguments{
\item{dir}{Directory in which IQ-TREE output files \code{.log} is stored.}
}
\value{
Numeric vector of likelihoods.
}
\description{
Reads likelihoods in respective \code{.log} files.
}
