% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_sc_network.R
\name{make_knn_networks}
\alias{make_knn_networks}
\title{make_knn_networks}
\usage{
make_knn_networks(
  obj,
  tree_list,
  k = NULL,
  k_dist = 4,
  use_branch_counts = T,
  pre_comp_path_dists_mat = NULL,
  edge_length_representations = c("min", "mean", "median")[1]
)
}
\arguments{
\item{obj}{Seurat object.}

\item{tree_list}{List of trees or \code{multiPhylo} object.}

\item{k}{Number of neighbors for the nearest neighbor graph construction. If \code{NULL} the maximal neighbor distance in \code{k_dist} is considered.}

\item{k_dist}{Maximal distance for the nearest neighbor graph construction, if \code{k = NULL}.}

\item{use_branch_counts}{If \code{TRUE} branch counts are used for the evaluation of the maximal distance \code{k_dist}, if \code{FALSE}, path lengths are used.}

\item{pre_comp_path_dists_mat}{Pre-computed distances output by \code{plot_tree_UMAP()} (optional).}

\item{edge_length_representations}{Summary statistic to compute representative edge length across nn-tree-graphs.}
}
\value{
Seurat object in which the network is stored as \code{igraph} object in \code{obj@misc$cellREST_igraph_obj}.
}
\description{
Constructs network given a list of trees by transforming trees into nnn-tree-graphs and uniting them.
}
