% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_plotting.R
\name{numerical_to_viridis}
\alias{numerical_to_viridis}
\title{numerical_to_viridis}
\usage{
numerical_to_viridis(x)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
A vector of colors from the color palette \code{viridis} of the same length of \code{x} mimicking numerical distances between entries of \code{x}.
}
\description{
numerical_to_viridis
}
