% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_sc_network.R
\name{plot_network}
\alias{plot_network}
\title{plot_network}
\usage{
plot_network(
  obj,
  network,
  umap_key = "network_umap",
  igraph_layout = FALSE,
  igraph_layout_type = c("layout_with_kk", "layout_with_fr")[1],
  layout_weight_attr = c("weight", "min_branch_count", "min_length")[1],
  edge_width_attr = c("weight", "min_branch_count", "min_length")[1],
  node_size = 6,
  edge_width_factor = 0.2,
  seed_graph_layout = 1,
  col_by = "cluster_names",
  col_pal = NULL,
  edge.color = NULL,
  show_plots = T,
  save_plots = F,
  pre = "",
  dir_plot = getwd(),
  im_width = 1,
  im_height = 1,
  im_res = 900
)
}
\arguments{
\item{obj}{Seurat object.}

\item{network}{\code{igraph} object.}

\item{umap_key}{Name of dimensionality reduction that is used for either plotting the network or for initializing a graph layout.}

\item{node_size}{Size of nodes.}

\item{edge_width_factor}{Scaling factor of edge widths.}

\item{seed_graph_layout}{Seed for computation of force-directed graph layout if \code{layout_by_weight = TRUE}.}

\item{col_by}{Name of metadata column to color cells by.}

\item{col_pal}{Color palette whose names match the unique values of \code{col_by}.}

\item{edge.color}{Vector of colors for edges.}

\item{show_plots}{boolean}

\item{save_plots}{boolean}

\item{pre}{Prefix used for name of created PNG.}

\item{dir_plot}{Directory in which PNG file is saved.}

\item{im_width}{Plot width.}

\item{im_height}{Plot height.}

\item{im_res}{Plot resolution.}

\item{layout_by_weight}{If \code{TRUE}, a force-directed graph layout is computed using network edge weights.}

\item{layout_model}{Either \code{layout_with_kk}, or \code{layout_with_fr} and is passed to \code{igraph} if \code{layout_by_weight = TRUE}.}
}
\description{
Plots \code{network} and saves plot as PNG.
}
