% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_visualize_scMLtrees.R
\name{plot_tree_UMAP}
\alias{plot_tree_UMAP}
\title{plot_tree_UMAP}
\usage{
plot_tree_UMAP(
  obj,
  tree_list,
  distance_measures = c("patristic_correlation", "RF", "wRF", "patristic", "nNodes")[1],
  save_patristic_dists = FALSE,
  png_list = NULL,
  lh_numbers = NULL,
  size_image = 0.1,
  umap_on_PCs = F,
  n_PCs = 50,
  min_dist_umap = 0.3,
  n_neighbors_umap = 30,
  spread_umap = 1,
  seed_umap = 42,
  show_plots = F,
  save_plots = F,
  pre = "",
  dir_plot = getwd(),
  im_width = 4,
  im_height = 4,
  im_res = 900,
  format = c("png", "pdf")[1],
  title = NULL
)
}
\arguments{
\item{obj}{Seurat object.}

\item{tree_list}{List of trees or \code{multiPhylo} object.}

\item{distance_measures}{Distance measure that is used to evaluate similarity between trees. Can be a measure compatible with \code{treespace::treespace()} or \code{patristic_correlation}.}

\item{save_patristic_dists}{If \code{TRUE} returns patristic distances per tree.}

\item{png_list}{List of paths to tree PNGs.}

\item{lh_numbers}{Vector of likelihoods that will specify a color code if \code{png_list = NULL}.}

\item{size_image}{Size of the PNGs in the tree-embedding visualization.}

\item{min_dist_umap}{Passed to \code{uwot::umap()}.}

\item{n_neighbors_umap}{Passed to \code{uwot::umap()}.}

\item{spread_umap}{Passed to \code{uwot::umap()}.}

\item{seed_umap}{Passed to \code{uwot::umap()}.}

\item{show_plots}{boolean}

\item{save_plots}{boolean}

\item{pre}{Prefix used for name of created plot file.}

\item{dir_plot}{Directory in which plot file is saved.}

\item{im_width}{Plot width.}

\item{im_height}{Plot height.}

\item{im_res}{Plot resolution.}

\item{format}{File format.}

\item{title}{Plot title.}
}
\description{
Plots and stores tree-embeddings using tree-to-tree similarities based on \code{distance_measures}.
}
