% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_sc_network.R
\name{prune_network}
\alias{prune_network}
\title{prune_network}
\usage{
prune_network(
  network,
  quantile = 0.9999,
  method = c("IQR", "left")[1],
  show_plots = T,
  save_plots = F,
  pre = "",
  dir_plot = getwd(),
  format = c("png", "pdf")[1],
  pt.size = 1,
  pt.alpha = 0.5,
  im_width_hist = 1.5,
  im_height_hist = 1.5,
  im_width_dot = 1.5,
  im_height_dot = 1.5,
  im_res = 900
)
}
\arguments{
\item{network}{\code{igraph} object.}

\item{quantile}{Theoretical quantile of the fitted normal distribution on logarithmized edge lengths.}

\item{method}{If \code{IQR} standard deviation is estimated based on the Interquantile Range, i.e. IQR/1.349. If \code{left} standard deviation is estimated using only values smaller than the median.}

\item{show_plots}{boolean}

\item{save_plots}{boolean}

\item{pre}{Prefix used for name of created plot files.}

\item{dir_plot}{Directory in which plot files is saved.}

\item{format}{File format.}

\item{pt.size}{Point size of edge length vs edge weight plot.}

\item{pt.alpha}{Point alpha of edge length vs edge weight plot.}

\item{im_width_hist}{Width of edge length histogram image.}

\item{im_height_hist}{Height of edge length histogram image.}

\item{im_width_dot}{Width of edge length vs edge weight plot image.}

\item{im_height_dot}{Height of edge length vs edge weight plot image.}

\item{im_res}{Plot resolution.}
}
\value{
Pruned network as \code{igraph} object.
}
\description{
Prunes edges of a network if their logarithmized lengths exceed the theoretical quantile of a fitted normal distribution
and visualizes edge length distribution as histogram and edge length vs edge weight plots.
}
