% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_import_from_IQ_Tree.R
\name{rate_categories}
\alias{rate_categories}
\title{rate_categories_from_iqtree}
\usage{
rate_categories(dir)
}
\arguments{
\item{dir}{Directory in which IQ-TREE output files \code{.iqtree} are stored.}
}
\value{
List of gamma shape alpha and rate vectors.
}
\description{
Reads rate categories from \code{.iqtree} files.
}
