% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_import_from_IQ_Tree.R
\name{rate_matrix_Q_from_iqtree}
\alias{rate_matrix_Q_from_iqtree}
\title{rate_matrix_from_iqtree}
\usage{
rate_matrix_Q_from_iqtree(dir)
}
\arguments{
\item{dir}{Directory in which IQ-TREE output files \code{.iqtree} are stored.}
}
\description{
Reads and plots rate matrix Q from \code{.iqtree} file as transition diagram.
}
