% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_import_from_IQ_Tree.R
\name{rate_parameter_R_from_log}
\alias{rate_parameter_R_from_log}
\title{rate_parameter_R_from_log}
\usage{
rate_parameter_R_from_log(dir)
}
\arguments{
\item{dir}{Directory in which IQ-TREE output files \code{.log} are stored.}
}
\description{
Reads and plots rate parameter R from \code{.log} as transition diagram.
}
