% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_pseudo_alignment.R
\name{seurat_to_fasta}
\alias{seurat_to_fasta}
\title{seurat_to_fasta}
\usage{
seurat_to_fasta(
  obj,
  varf = NULL,
  pre = unique(obj$orig.ident)[1],
  binning_method = c("ACGT")[1],
  assay_in = DefaultAssay(obj),
  dir = getwd()
)
}
\arguments{
\item{obj}{Seurat object.}

\item{varf}{Character vector of highly variable genes names.}

\item{pre}{Prefix for the FASTA file name.}

\item{binning_method}{Passed to function \code{binning()}.}

\item{assay_in}{Name of Seurat object assay to use.}

\item{dir}{Directory for storing FASTA file.}
}
\description{
Creates and writes a pseudo-alignment from normalized counts (stored in \code{obj@assays[[assay_in]]$data}). Genes are subset to genes specified in \code{varf}.
}
