% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_plotting.R
\name{theme_minimal_plots}
\alias{theme_minimal_plots}
\title{theme_minimal_plots}
\usage{
theme_minimal_plots(text_size = 6, text_col = "black", ...)
}
\arguments{
\item{text_size}{Text Size.}

\item{text_col}{Text Color.}

\item{...}{Additional parameters passed to \code{ggplot2::theme()}.}
}
\description{
Adapts legend ticks length and plot title position and fixes text size, text color, and a transparent background complementing \code{ggplot2::theme_minimal()}.
}
