% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellREST_visualize_scMLtrees.R
\name{tree_list_pngs}
\alias{tree_list_pngs}
\title{tree_list_pngs}
\usage{
tree_list_pngs(
  obj,
  tree_list,
  col_by = "cluster_names",
  col_pal = NULL,
  plot_OTU = F,
  verbose = T,
  branch_width = 1,
  plot_tip_as_point = T,
  point_size = 0.5,
  check_for_existing_PNGs = T,
  pre = "",
  dir_plots = getwd(),
  im_width = 3,
  im_height = 2.5,
  im_res = 900
)
}
\arguments{
\item{obj}{Seurat object.}

\item{tree_list}{List of trees or \code{multiPhylo} object.}

\item{col_by}{Name of metadata column to color cells by.}

\item{col_pal}{Color palette whose names match the unique values of \code{col_by}.}

\item{plot_OTU}{Color trees using \code{ggtree::groupOTU()}.}

\item{verbose}{boolean}

\item{branch_width}{Factor of branch widths.}

\item{plot_tip_as_point}{boolean}

\item{point_size}{Tip point size.}

\item{check_for_existing_PNGs}{boolean}

\item{pre}{Prefix used for name of created plot file.}

\item{im_width}{Plot width.}

\item{im_height}{Plot height.}

\item{im_res}{Plot resolution.}

\item{dir_plot}{Directory in which plot file is saved.}
}
\description{
Creates and stores tree plots as PNGs.
}
